📂 backend restructure
