+++ backend/.dockerignore
@@ -1,0 +1,3 @@
+.venv
+dist
+.coverage
