+++ backend/packages/backend/backend/__init__.py
@@ -2,6 +2,6 @@
 camply-backend
 """
 
-from camply_backend.__about__ import __version__
+from backend.__about__ import __version__
 
 __all__ = ["__version__"]
