+++ backend/pyproject.toml
@@ -2,53 +2,14 @@
 name = "camply-backend"
 version = "0.2.0"
 description = "A full-stack application to help you find campsites at sold-out campgrounds"
-readme = "README.md"
-authors = [
-  {name = "Justin Flannery", email = "justin.flannery@juftin.com"}
-]
 requires-python = ">=3.9,<4.0"
-dependencies = [
-  "fastapi>=0.115.12",
-  "gunicorn>=23.0.0",
-  "importlib-metadata>=6.5.0; python_version < '3.10'",
-  "orjson>=3.10.16",
-  "pydantic>=2.10.6",
-  "pydantic-settings>=2.8.1",
-  "uvicorn-worker>=0.3.0",
-  "uvicorn[standard]>=0.34.0"
-]
+dependencies = ["backend"]
 
-[project.scripts]
-camply-backend-debug = "camply_backend.app:main"
+[tool.uv.sources]
+backend = { workspace = true }
 
-[dependency-groups]
-dev = [
-  "httpx>=0.28.1",
-  "pytest>=8.3.5",
-  "pytest-cov>=6.0.0",
-  "ruff>=0.11.2",
-  "mypy>=1.15.0",  # mypy --install-types
-  "pip>=24.3.1",
-  "types-pyyaml>=6.0.12.20250516",
-  "types-ujson>=5.10.0.20250326"
-]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["camply_backend"]
-
-[tool.pytest.ini_options]
-addopts = "-v --cov=camply_backend --cov-report=term-missing"
-
-[tool.coverage.report]
-exclude_lines = [
-  "no cov",
-  "if __name__ == .__main__.:",
-  "if TYPE_CHECKING:"
-]
+[tool.uv.workspace]
+members = ["packages/*"]
 
 [tool.ruff]
 target-version = "py39"
@@ -83,7 +44,7 @@ known-first-party = ["camply_backend"]
 
 [tool.ruff.lint.per-file-ignores]
 # Tests can use magic values, assertions, and relative imports
-"tests/**/*" = ["PLR2004", "S101", "TID252"]
+"**/tests/**/*" = ["PLR2004", "S101", "TID252"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "numpy"
